Creation of soft fonts for TrueType fonts in PCL
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/Temp_PCLSoftFonts@1685112 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/truetype/OFDirTabEntry.java
@@ -30,7 +30,7 @@ import java.util.Arrays;
 public class OFDirTabEntry {
 
     private byte[] tag = new byte[4];
-    private int checksum;
+    private long checksum;
     private long offset;
     private long length;
 
@@ -74,7 +74,7 @@ public class OFDirTabEntry {
      * Returns the checksum.
      * @return int
      */
-    public int getChecksum() {
+    public long getChecksum() {
         return checksum;
     }
